Refactor JSON tests of OrderedMap
