+++ pkg/orderedmap/orderedmap_test.go
@@ -39,15 +39,15 @@ func TestOrderedMap(t *testing.T) {
 		t.Error("Set strings second index")
 	}
 	// mixed slice
-	o.Set("mixed", []any{
+	o.Set("mixed", []interface{}{
 		1,
 		"1",
 	})
 	v, _ = o.Get("mixed")
-	if v.([]any)[0].(int) != 1 {
+	if v.([]interface{})[0].(int) != 1 {
 		t.Error("Set mixed int")
 	}
-	if v.([]any)[1].(string) != "1" {
+	if v.([]interface{})[1].(string) != "1" {
 		t.Error("Set mixed string")
 	}
 	// overriding existing key
@@ -82,393 +82,6 @@ func TestOrderedMap(t *testing.T) {
 	}
 	if _, ok := o.Get("strings"); ok {
 		t.Error("Delete did not remove 'strings' key")
-	}
-}
-
-func TestBlankMarshalJSON(t *testing.T) {
-	t.Parallel()
-	o := New()
-	// blank map
-	b, err := json.Marshal(o)
-	if err != nil {
-		t.Error("Marshalling blank map to json", err)
-	}
-
-	// check json is correctly ordered
-	if s := string(b); s != `{}` {
-		t.Error("JSON Marshaling blank map value is incorrect", s)
-	}
-	// convert to indented json
-	bi, err := json.MarshalIndent(o, "", "  ")
-	if err != nil {
-		t.Error("Marshalling indented json for blank map", err)
-	}
-	si := string(bi)
-	ei := `{}`
-	if si != ei {
-		t.Error("JSON MarshalIndent blank map value is incorrect", si)
-	}
-}
-
-func TestMarshalJSON(t *testing.T) {
-	t.Parallel()
-	o := New()
-	// number
-	o.Set("number", 3)
-	// string
-	o.Set("string", "x")
-	// string
-	o.Set("specialstring", "\\.<>[]{}_-")
-	// new value keeps key in old position
-	o.Set("number", 4)
-	// keys not sorted alphabetically
-	o.Set("z", 1)
-	o.Set("a", 2)
-	o.Set("b", 3)
-	// slice
-	o.Set("slice", []any{
-		"1",
-		1,
-	})
-	// orderedmap
-	v := New()
-	v.Set("e", 1)
-	v.Set("a", 2)
-	o.Set("orderedmap", v)
-	// escape key
-	o.Set("test\n\r\t\\\"ing", 9)
-	// convert to json
-	b, err := json.Marshal(o)
-	if err != nil {
-		t.Error("Marshalling json", err)
-	}
-	// check json is correctly ordered
-	if s := string(b); s != `{"number":4,"string":"x","specialstring":"\\.\u003c\u003e[]{}_-","z":1,"a":2,"b":3,"slice":["1",1],"orderedmap":{"e":1,"a":2},"test\n\r\t\\\"ing":9}` {
-		t.Error("JSON Marshal value is incorrect", s)
-	}
-	// convert to indented json
-	bi, err := json.MarshalIndent(o, "", "  ")
-	if err != nil {
-		t.Error("Marshalling indented json", err)
-	}
-	si := string(bi)
-	ei := `{
-  "number": 4,
-  "string": "x",
-  "specialstring": "\\.\u003c\u003e[]{}_-",
-  "z": 1,
-  "a": 2,
-  "b": 3,
-  "slice": [
-    "1",
-    1
-  ],
-  "orderedmap": {
-    "e": 1,
-    "a": 2
-  },
-  "test\n\r\t\\\"ing": 9
-}`
-	if si != ei {
-		t.Error("JSON MarshalIndent value is incorrect", si)
-	}
-}
-
-func TestUnmarshalJSON(t *testing.T) {
-	t.Parallel()
-	s := `{
-  "number": 4,
-  "string": "x",
-  "z": 1,
-  "a": "should not break with unclosed { character in value",
-  "b": 3,
-  "slice": [
-    "1",
-    1
-  ],
-  "orderedmap": {
-    "e": 1,
-    "a { nested key with brace": "with a }}}} }} {{{ brace value",
-	"after": {
-		"link": "test {{{ with even deeper nested braces }"
-	}
-  },
-  "test\"ing": 9,
-  "after": 1,
-  "multitype_array": [
-    "test",
-	1,
-	{ "map": "obj", "it" : 5, ":colon in key": "colon: in value" },
-	[{"inner": "map"}]
-  ],
-  "should not break with { character in key": 1
-}`
-	o := New()
-	err := json.Unmarshal([]byte(s), &o)
-	if err != nil {
-		t.Error("JSON Unmarshal error", err)
-	}
-	// Check the root keys
-	expectedKeys := []string{
-		"number",
-		"string",
-		"z",
-		"a",
-		"b",
-		"slice",
-		"orderedmap",
-		"test\"ing",
-		"after",
-		"multitype_array",
-		"should not break with { character in key",
-	}
-	k := o.Keys()
-	for i := range k {
-		if k[i] != expectedKeys[i] {
-			t.Error("Unmarshal root key order", i, k[i], "!=", expectedKeys[i])
-		}
-	}
-	// Check nested maps are converted to orderedmaps
-	// nested 1 level deep
-	expectedKeys = []string{
-		"e",
-		"a { nested key with brace",
-		"after",
-	}
-	vi, ok := o.Get("orderedmap")
-	if !ok {
-		t.Error("Missing key for nested map 1 deep")
-	}
-	v := vi.(*OrderedMap)
-	k = v.Keys()
-	for i := range k {
-		if k[i] != expectedKeys[i] {
-			t.Error("Key order for nested map 1 deep ", i, k[i], "!=", expectedKeys[i])
-		}
-	}
-	// nested 2 levels deep
-	expectedKeys = []string{
-		"link",
-	}
-	vi, ok = v.Get("after")
-	if !ok {
-		t.Error("Missing key for nested map 2 deep")
-	}
-	v = vi.(*OrderedMap)
-	k = v.Keys()
-	for i := range k {
-		if k[i] != expectedKeys[i] {
-			t.Error("Key order for nested map 2 deep", i, k[i], "!=", expectedKeys[i])
-		}
-	}
-	// multitype array
-	expectedKeys = []string{
-		"map",
-		"it",
-		":colon in key",
-	}
-	vislice, ok := o.Get("multitype_array")
-	if !ok {
-		t.Error("Missing key for multitype array")
-	}
-	vslice := vislice.([]any)
-	vmap := vslice[2].(*OrderedMap)
-	k = vmap.Keys()
-	for i := range k {
-		if k[i] != expectedKeys[i] {
-			t.Error("Key order for nested map 2 deep", i, k[i], "!=", expectedKeys[i])
-		}
-	}
-	// nested map 3 deep
-	vislice, _ = o.Get("multitype_array")
-	vslice = vislice.([]any)
-	expectedKeys = []string{"inner"}
-	vinnerslice := vslice[3].([]any)
-	vinnermap := vinnerslice[0].(*OrderedMap)
-	k = vinnermap.Keys()
-	for i := range k {
-		if k[i] != expectedKeys[i] {
-			t.Error("Key order for nested map 3 deep", i, k[i], "!=", expectedKeys[i])
-		}
-	}
-}
-
-func TestUnmarshalJSONDuplicateKeys(t *testing.T) {
-	t.Parallel()
-	s := `{
-		"a": [{}, []],
-		"b": {"x":[1]},
-		"c": "x",
-		"d": {"x":1},
-		"b": [{"x":[]}],
-		"c": 1,
-		"d": {"y": 2},
-		"e": [{"x":1}],
-		"e": [[]],
-		"e": [{"z":2}],
-		"a": {},
-		"b": [[1]]
-	}`
-	o := New()
-	err := json.Unmarshal([]byte(s), &o)
-	if err != nil {
-		t.Error("JSON Unmarshal error with special chars", err)
-	}
-	expectedKeys := []string{
-		"c",
-		"d",
-		"e",
-		"a",
-		"b",
-	}
-	keys := o.Keys()
-	if len(keys) != len(expectedKeys) {
-		t.Error("Unmarshal key count", len(keys), "!=", len(expectedKeys))
-	}
-	for i, key := range keys {
-		if key != expectedKeys[i] {
-			t.Errorf("Unmarshal root key order: %d, %q != %q", i, key, expectedKeys[i])
-		}
-	}
-	vimap, _ := o.Get("a")
-	_ = vimap.(*OrderedMap)
-	vislice, _ := o.Get("b")
-	_ = vislice.([]any)
-	vival, _ := o.Get("c")
-	_ = vival.(float64)
-
-	vimap, _ = o.Get("d")
-	m := vimap.(*OrderedMap)
-	expectedKeys = []string{"y"}
-	keys = m.Keys()
-	if len(keys) != len(expectedKeys) {
-		t.Error("Unmarshal key count", len(keys), "!=", len(expectedKeys))
-	}
-	for i, key := range keys {
-		if key != expectedKeys[i] {
-			t.Errorf("Unmarshal key order: %d, %q != %q", i, key, expectedKeys[i])
-		}
-	}
-
-	vislice, _ = o.Get("e")
-	m = vislice.([]any)[0].(*OrderedMap)
-	expectedKeys = []string{"z"}
-	keys = m.Keys()
-	if len(keys) != len(expectedKeys) {
-		t.Error("Unmarshal key count", len(keys), "!=", len(expectedKeys))
-	}
-	for i, key := range keys {
-		if key != expectedKeys[i] {
-			t.Errorf("Unmarshal key order: %d, %q != %q", i, key, expectedKeys[i])
-		}
-	}
-}
-
-func TestUnmarshalJSONSpecialChars(t *testing.T) {
-	t.Parallel()
-	s := `{ " \u0041\n\r\t\\\\\\\\\\\\ "  : { "\\\\\\" : "\\\\\"\\" }, "\\":  " \\\\ test ", "\n": "\r" }`
-	o := New()
-	err := json.Unmarshal([]byte(s), &o)
-	if err != nil {
-		t.Error("JSON Unmarshal error with special chars", err)
-	}
-	expectedKeys := []string{
-		" \u0041\n\r\t\\\\\\\\\\\\ ",
-		"\\",
-		"\n",
-	}
-	keys := o.Keys()
-	if len(keys) != len(expectedKeys) {
-		t.Error("Unmarshal key count", len(keys), "!=", len(expectedKeys))
-	}
-	for i, key := range keys {
-		if key != expectedKeys[i] {
-			t.Errorf("Unmarshal root key order: %d, %q != %q", i, key, expectedKeys[i])
-		}
-	}
-}
-
-func TestUnmarshalJSONArrayOfMaps(t *testing.T) {
-	t.Parallel()
-	s := `
-{
-  "name": "test",
-  "percent": 6,
-  "breakdown": [
-    {
-      "name": "a",
-      "percent": 0.9
-    },
-    {
-      "name": "b",
-      "percent": 0.9
-    },
-    {
-      "name": "d",
-      "percent": 0.4
-    },
-    {
-      "name": "e",
-      "percent": 2.7
-    }
-  ]
-}
-`
-	o := New()
-	err := json.Unmarshal([]byte(s), &o)
-	if err != nil {
-		t.Error("JSON Unmarshal error", err)
-	}
-	// Check the root keys
-	expectedKeys := []string{
-		"name",
-		"percent",
-		"breakdown",
-	}
-	k := o.Keys()
-	for i := range k {
-		if k[i] != expectedKeys[i] {
-			t.Error("Unmarshal root key order", i, k[i], "!=", expectedKeys[i])
-		}
-	}
-	// Check nested maps are converted to orderedmaps
-	// nested 1 level deep
-	expectedKeys = []string{
-		"name",
-		"percent",
-	}
-	vi, ok := o.Get("breakdown")
-	if !ok {
-		t.Error("Missing key for nested map 1 deep")
-	}
-	vs := vi.([]any)
-	for _, vInterface := range vs {
-		v := vInterface.(*OrderedMap)
-		k = v.Keys()
-		for i := range k {
-			if k[i] != expectedKeys[i] {
-				t.Error("Key order for nested map 1 deep ", i, k[i], "!=", expectedKeys[i])
-			}
-		}
-	}
-}
-
-func TestUnmarshalJSONStruct(t *testing.T) {
-	t.Parallel()
-	var v struct {
-		Data *OrderedMap `json:"data"`
-	}
-
-	err := json.Unmarshal([]byte(`{ "data": { "x": 1 } }`), &v)
-	if err != nil {
-		t.Fatalf("JSON unmarshal error: %v", err)
-	}
-
-	x, ok := v.Data.Get("x")
-	if !ok {
-		t.Errorf("missing expected key")
-	} else if x != float64(1) {
-		t.Errorf("unexpected value: %#v", x)
 	}
 }
 
@@ -599,9 +212,9 @@ func TestOrderedMapGetNested(t *testing.T) {
 	root := New()
 	nested := New()
 	nested.Set(`key`, `value`)
-	nested.Set(`slice`, []any{1, 2, 3})
+	nested.Set(`slice`, []interface{}{1, 2, 3})
 	root.Set(`nested`, nested)
-	root.Set(`slice`, []any{1, 2, 3})
+	root.Set(`slice`, []interface{}{1, 2, 3})
 
 	// Missing root map key
 	value, found, err := root.GetNested(`foo`)
